I moved it to content folder
