+++ src/feature/Content/Pages.js
@@ -1,6 +1,6 @@
 import { useDispatch, useSelector } from "react-redux";
-import { fetchMovies, selectMovies, selectMoviesStatus } from "../moviesSlice";
-import { Content } from "../Content";
+import { fetchMovies, selectMovies, selectMoviesStatus } from "../Movie/moviesSlice";
+import { Content } from "./Content";
 import { useEffect } from "react";
 
 export const Pages = () => {
